mkiats-dev-deployment/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
anyio==4.8.0
attrs==25.1.0
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==4.9.0
azure-functions==1.21.3
azure-functions-durable==1.2.10
azure-identity==1.19.0
azure-mgmt-core==1.5.0
azure-mgmt-resource==23.2.0
azure-mgmt-storage==21.2.1
azure-mgmt-web==7.3.1
azure-storage-blob==12.24.1
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
frozenlist==1.5.0
furl==2.1.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
orderedmultidict==1.0.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
portalocker==2.10.1
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
